pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Manh Pham <manhpham90vn@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic = {extras = ["email"], version = "^2.10.3"}
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
cryptography = "^44.0.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
